I have added the file for 7*

+++ 0x0F-python-object_relational_mapping/model_state.py
@@ -17,6 +17,6 @@ class State(Base):
     name (sqlalchemy.String): name of state
     """
     __tablename__ = 'states'
-    id = Column(Integer(11), primary_key=True,
+    id = Column(Integer, primary_key=True,
                 autoincrement=True, nullable=False)
     name = Column(String(128), nullable=False)
